requirements.txt
# Core dependencies
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
langchain-core==0.2.43
openai==1.40.8

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# UI
streamlit==1.31.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Development tools (opcional, pero incluidos para facilitar setup)
black==23.12.1
isort==5.13.2
pytest==7.4.4
pytest-cov==4.1.0
mypy==1.8.0
pre-commit==3.6.0
